+++ src/main/java/com/project/doongdoong/domain/voice/domain/FileExtension.java
@@ -1,4 +1,4 @@
-package com.project.doongdoong.domain.voice.model;
+package com.project.doongdoong.domain.voice.domain;
 
 import com.project.doongdoong.domain.voice.exception.NotFoundFileExtension;
 import lombok.Getter;
