+++ src/main/java/com/project/doongdoong/global/util/WebClientUtil.java
@@ -2,7 +2,7 @@ package com.project.doongdoong.global.util;
 
 
 import com.project.doongdoong.domain.analysis.adapter.in.dto.FellingStateCreateResponse;
-import com.project.doongdoong.domain.voice.model.Voice;
+import com.project.doongdoong.domain.voice.domain.Voice;
 import com.project.doongdoong.global.dto.request.ConsultRequest;
 import com.project.doongdoong.global.dto.request.VoiceToS3Request;
 import com.project.doongdoong.global.dto.response.CounselAiResponse;
