+++ src/test/java/com/project/doongdoong/domain/voice/service/VoiceServiceImpTest.java
@@ -5,9 +5,10 @@ import com.amazonaws.services.s3.model.DeleteObjectsRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectResult;
 import com.project.doongdoong.domain.question.domain.QuestionContent;
-import com.project.doongdoong.domain.voice.dto.response.VoiceDetailResponseDto;
-import com.project.doongdoong.domain.voice.model.Voice;
-import com.project.doongdoong.domain.voice.repository.VoiceRepository;
+import com.project.doongdoong.domain.voice.adapter.in.dto.VoiceDetailResponseDto;
+import com.project.doongdoong.domain.voice.application.VoiceServiceImp;
+import com.project.doongdoong.domain.voice.domain.Voice;
+import com.project.doongdoong.domain.voice.application.port.out.VoiceRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
