+++ src/test/java/com/project/doongdoong/module/IntegrationSupportTest.java
@@ -1,6 +1,5 @@
 package com.project.doongdoong.module;
 
-import com.project.doongdoong.domain.voice.service.VoiceService;
 import com.project.doongdoong.global.util.WebClientUtil;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
